move status checking to lib & call from both commands
